# docs/openapi.json
{
  "components": {
    "schemas": {
      "AuditVaultResponse": {
        "properties": {
          "broken_links": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Broken Links",
            "type": "array"
          },
          "stale_documents": {
            "items": {
              "type": "string"
            },
            "title": "Stale Documents",
            "type": "array"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "suggested_actions": {
            "items": {
              "type": "string"
            },
            "title": "Suggested Actions",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "summary"
        ],
        "title": "AuditVaultResponse",
        "type": "object"
      },
...
      "CipherAuditApiRequest": {
        "properties": {
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "deps": {
            "$ref": "#/components/schemas/CipherDepsModel"
          },
          "mode": {
            "default": "audit",
            "enum": [
              "audit",
              "repair"
            ],
            "title": "Mode",
            "type": "string"
          },
          "op": {
            "const": "audit_vault",
            "default": "audit_vault",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "deps"
        ],
        "title": "CipherAuditApiRequest",
        "type": "object"
      },
      "CipherChunkingApiRequest": {
        "properties": {
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "model_info": {
            "title": "Model Info",
            "type": "string"
          },
          "op": {
            "const": "suggest_chunking_strategy",
            "default": "suggest_chunking_strategy",
            "title": "Op",
            "type": "string"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "model_info"
        ],
        "title": "CipherChunkingApiRequest",
        "type": "object"
      },
...
      "CipherExplainApiRequest": {
        "properties": {
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "doc_a": {
            "title": "Doc A",
            "type": "string"
          },
          "doc_b": {
            "title": "Doc B",
            "type": "string"
          },
          "op": {
            "const": "explain_connection",
            "default": "explain_connection",
            "title": "Op",
            "type": "string"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "doc_a",
          "doc_b"
        ],
        "title": "CipherExplainApiRequest",
        "type": "object"
      },
      "CipherStubReviewApiRequest": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/StubCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "op": {
            "const": "review_stub_candidates",
            "default": "review_stub_candidates",
            "title": "Op",
            "type": "string"
          },
          "stub_dir": {
            "default": "entities",
            "title": "Stub Dir",
            "type": "string"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "title": "CipherStubReviewApiRequest",
        "type": "object"
      },
      "DiscoverRequest": {
        "properties": {
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "db_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db Path"
          },
          "max_total": {
            "default": 10,
            "title": "Max Total",
            "type": "integer"
          },
          "threshold": {
            "default": 0.2,
            "title": "Threshold",
            "type": "number"
          },
          "top_k": {
            "default": 3,
            "title": "Top K",
            "type": "integer"
          }
        },
        "title": "DiscoverRequest",
        "type": "object"
      },
      "DiscoverResponse": {
        "properties": {
          "database_path": {
            "title": "Database Path",
            "type": "string"
          },
          "discoveries": {
            "items": {
              "$ref": "#/components/schemas/DiscoveryResult"
            },
            "title": "Discoveries",
            "type": "array"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "database_path",
          "threshold"
        ],
        "title": "DiscoverResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "dependency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency"
          },
          "error_type": {
            "title": "Error Type",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "error_type",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExplainConnectionResponse": {
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "explanation"
        ],
        "title": "ExplainConnectionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewStubCandidatesResponse": {
        "properties": {
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/StubCandidateReview"
            },
            "title": "Reviews",
            "type": "array"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ReviewStubCandidatesResponse",
        "type": "object"
      },
...
      "SuggestChunkingStrategyResponse": {
        "properties": {
          "max_chunk_size": {
            "title": "Max Chunk Size",
            "type": "integer"
          },
          "min_chunk_size": {
            "title": "Min Chunk Size",
            "type": "integer"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "max_chunk_size",
          "min_chunk_size",
          "rationale"
        ],
        "title": "SuggestChunkingStrategyResponse",
        "type": "object"
      },
...
      "ValidateResponse": {
        "properties": {
          "broken_links": {
            "items": {
              "$ref": "#/components/schemas/BrokenLinkResult"
            },
            "title": "Broken Links",
            "type": "array"
          },
          "database_path": {
            "title": "Database Path",
            "type": "string"
          }
        },
        "required": [
          "database_path"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Expose Synapse indexing, search, discovery, validation, and Cipher operations over JSON/HTTP for PWAs and other web clients.",
    "summary": "HTTP and OpenAPI interface for Synapse retrieval and Cipher reasoning.",
    "title": "Synapse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cipher/audit": {
      "post": {
        "operationId": "post_cipher_audit_cipher_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CipherAuditApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVaultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Cipher Audit",
        "tags": [
          "cipher"
        ]
      }
    },
    "/cipher/chunking-strategy": {
      "post": {
        "operationId": "post_cipher_chunking_strategy_cipher_chunking_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CipherChunkingApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestChunkingStrategyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Post Cipher Chunking Strategy",
        "tags": [
          "cipher"
        ]
      }
    },
    "/cipher/explain": {
      "post": {
        "operationId": "post_cipher_explain_cipher_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CipherExplainApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainConnectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Post Cipher Explain",
        "tags": [
          "cipher"
        ]
      }
    },
...
    "/cipher/review-stubs": {
      "post": {
        "operationId": "post_cipher_review_stubs_cipher_review_stubs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CipherStubReviewApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStubCandidatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Post Cipher Review Stubs",
        "tags": [
          "cipher"
        ]
      }
    },
    "/discover": {
      "post": {
        "operationId": "post_discover_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Discover",
        "tags": [
          "synapse"
        ]
      }
    },
...
    "/index": {
      "post": {
        "operationId": "post_index_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Index",
        "tags": [
          "synapse"
        ]
      }
    },
    "/search": {
      "post": {
        "operationId": "post_search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "424": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed Dependency"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Search",
        "tags": [
          "synapse"
        ]
      }
    },
    "/validate": {
      "post": {
        "operationId": "post_validate_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Validate",
        "tags": [
          "synapse"
        ]
      }
    }
  }
}